+++ partiql-lang/src/main/kotlin/org/partiql/lang/planner/transforms/AstToLogicalVisitorTransform.kt
@@ -8,10 +8,14 @@ import org.partiql.lang.domains.PartiqlAst
 import org.partiql.lang.domains.PartiqlAstToPartiqlLogicalVisitorTransform
 import org.partiql.lang.domains.PartiqlLogical
 import org.partiql.lang.domains.metaContainerOf
+import org.partiql.lang.errors.ErrorCode
 import org.partiql.lang.errors.Problem
 import org.partiql.lang.errors.ProblemHandler
+import org.partiql.lang.eval.EvaluationSession
 import org.partiql.lang.eval.builtins.CollectionAggregationFunction
 import org.partiql.lang.eval.builtins.ExprFunctionCurrentUser
+import org.partiql.lang.eval.err
+import org.partiql.lang.eval.errorContextFrom
 import org.partiql.lang.eval.physical.sourceLocationMetaOrUnknown
 import org.partiql.lang.eval.visitors.VisitorTransformBase
 import org.partiql.lang.planner.PlanningProblemDetails
@@ -82,9 +86,18 @@ internal class AstToLogicalVisitorTransform(
         }
     }
 
-    override fun transformExprCurrentUser(node: PartiqlAst.Expr.CurrentUser): PartiqlLogical.Expr = PartiqlLogical.build {
+    override fun transformExprSessionAttribute(node: PartiqlAst.Expr.SessionAttribute): PartiqlLogical.Expr.Call = PartiqlLogical.build {
+        val functionName = when (node.value.text.toUpperCase()) {
+            EvaluationSession.Constants.CURRENT_USER_KEY -> ExprFunctionCurrentUser.FUNCTION_NAME
+            else -> err(
+                "Unsupported session attribute: ${node.value.text}",
+                errorCode = ErrorCode.SEMANTIC_PROBLEM,
+                errorContext = errorContextFrom(node.metas),
+                internal = false
+            )
+        }
         call(
-            funcName = ExprFunctionCurrentUser.NAME,
+            funcName = functionName,
             args = emptyList()
         )
     }